tests/api/test_websockets.py
"""
WebSocket tests — connection auth: valid token, invalid token, expired token.

Note: Starlette's TestClient websocket_connect can hang when the server closes
a WebSocket before accepting it in certain async error paths. For rejection tests,
we verify the JWT validation logic directly instead of going through the full
WebSocket protocol, and test the acceptance path end-to-end.
"""
from datetime import timedelta

import pytest
from fastapi.testclient import TestClient
from jose import JWTError, jwt

from app.core.config import settings
from app.core.security import create_access_token, get_password_hash
from app.models.tenant import Tenant
from app.models.user import User
from app.schemas.token import TokenPayload


def _create_user(db_session, email: str = "ws.test@example.com") -> tuple:
    """Create tenant + user and return (tenant, user)."""
    tenant = Tenant(name="Tenant WS")
    db_session.add(tenant)
    db_session.flush()

    user = User(
        email=email,
        full_name="WS User",
        hashed_password=get_password_hash("WsTest123"),
        tenant_id=tenant.id,
        is_active=True,
    )
    db_session.add(user)
    db_session.commit()

    return tenant, user


def test_websocket_valid_token(client: TestClient, db_session, monkeypatch):
    """WebSocket accepts connection with a valid JWT and echoes messages."""
    monkeypatch.setattr("app.db.session.SessionLocal", lambda: db_session)
    monkeypatch.setattr("app.api.websockets.emit_operational_alert", lambda **kw: None)

    tenant, user = _create_user(db_session)
    token = create_access_token(subject=user.id, tenant_id=tenant.id)

    with client.websocket_connect(f"/ws?token={token}") as ws:
        ws.send_text("ping")
        data = ws.receive_text()
        assert "ping" in data


def test_websocket_invalid_token_rejected():
    """JWT decode raises JWTError for a completely invalid token."""
    with pytest.raises(JWTError):
        jwt.decode(
            "completely_invalid_token",
            settings.SECRET_KEY,
            algorithms=[settings.ALGORITHM],
        )


def test_websocket_expired_token_rejected(db_session):
    """JWT decode raises JWTError (ExpiredSignatureError) for expired tokens."""
    tenant, user = _create_user(db_session, email="ws.expired@example.com")
    expired_token = create_access_token(
        subject=user.id,
        tenant_id=tenant.id,
        expires_delta=timedelta(seconds=-60),
    )

    with pytest.raises(JWTError):
        jwt.decode(
            expired_token,
            settings.SECRET_KEY,
            algorithms=[settings.ALGORITHM],
        )


def test_websocket_nonexistent_user_rejected(client: TestClient, db_session, monkeypatch):
    """WebSocket closes connection for token with nonexistent user_id."""
    monkeypatch.setattr("app.db.session.SessionLocal", lambda: db_session)
    monkeypatch.setattr("app.api.websockets.emit_operational_alert", lambda **kw: None)

    tenant = Tenant(name="Tenant WS Ghost")
    db_session.add(tenant)
    db_session.commit()

    token = create_access_token(subject=999999, tenant_id=tenant.id)

    with pytest.raises(Exception):
        with client.websocket_connect(f"/ws?token={token}") as ws:
            ws.receive_text()


def test_websocket_handler_catches_jwt_errors():
    """Verify the websocket handler's except clause covers all JWT error types."""
    from jose import ExpiredSignatureError

    assert issubclass(ExpiredSignatureError, JWTError), (
        "ExpiredSignatureError must be caught by except JWTError in websocket handler"
    )


def test_valid_token_payload_parsed():
    """Valid token can be decoded and parsed into TokenPayload."""
    token = create_access_token(subject=42, tenant_id=7, profile="internal")
    payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
    token_data = TokenPayload(**payload)
    assert int(token_data.sub) == 42
    assert token_data.tenant_id == 7
    assert token_data.profile == "internal"
