requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
email-validator
sqlalchemy
sqlalchemy-utils
alembic
psycopg2-binary
redis
celery
python-multipart
python-jose[cryptography]
passlib[bcrypt]
boto3
GeoAlchemy2
fpdf2
litellm
mempalace>=3.0.0
httpx
slowapi>=0.1.9
pytest
pytest-cov
testcontainers[postgres]
ruff
mypy
